requirements.txt
PyYAML>=6.0.1
typer>=0.9.0
rich>=13.7.0
pydantic>=2.5.0
pytest>=7.4.0
click>=8.0.0
typing-extensions>=4.8.0
pathlib>=1.0.1
